requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Migrate
email-validator
python-dotenv
google-genai
gunicorn
razorpay